# api/docs/swagger.yaml
definitions:
  model.Day:
    properties:
      conditions:
        type: string
      datetime:
        type: string
      datetimeEpoch:
        type: integer
      hours:
        items:
          $ref: '#/definitions/model.HourlyData'
        type: array
      humidity:
        type: number
      temp:
        type: number
      tempmax:
        type: number
      tempmin:
        type: number
    type: object
  model.GetAllWeatherResponse:
    properties:
      total_count:
        type: integer
      weathers:
        items:
          $ref: '#/definitions/model.Weather'
        type: array
    type: object
  model.HourlyData:
    properties:
      datetime:
        type: string
      datetimeEpoch:
        type: integer
      humidity:
        type: number
      icon:
        type: string
      temp:
        type: number
    type: object
  model.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
  model.Weather:
    properties:
      day:
        properties:
          condition:
            type: string
          humidity:
            type: number
          temperature:
            type: number
          temperature_max:
            type: number
          temperature_min:
            type: number
          time:
            type: string
        type: object
      hour:
        properties:
          condition:
            type: string
          humidity:
            type: number
          temperature:
            type: number
          time:
            type: string
        type: object
      location:
        properties:
          country:
            type: string
          id:
            type: string
          name:
            type: string
        type: object
    type: object
  model.WeatherResponse:
    properties:
      address:
        type: string
      days:
        items:
          $ref: '#/definitions/model.Day'
        type: array
      hours:
        items:
          $ref: '#/definitions/model.HourlyData'
        type: array
      resolvedAddress:
        type: string
      timezone:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
  /weather:
    get:
      consumes:
      - application/json
      description: This API gets weather data by country name and returns its info
      parameters:
      - description: city_name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WeatherResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Get weather data by country name
      tags:
      - weather
  /weatherget:
    get:
      consumes:
      - application/json
      description: This API gets all weather data with optional filters
      parameters:
      - description: City name
        in: query
        name: city
        type: string
      - description: Weather condition
        in: query
        name: condition
        type: string
      - description: Minimum temperature
        in: query
        name: min_temp
        type: number
      - description: Maximum temperature
        in: query
        name: max_temp
        type: number
      - description: Start time (YYYY-MM-DD)
        in: query
        name: start_time
        type: string
      - description: End time (YYYY-MM-DD)
        in: query
        name: end_time
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetAllWeatherResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all weather data with filters
      tags:
      - weather
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"